fix: calibration data types and use calibrated values

+++ code/arduino-tct/constants.hpp
@@ -15,9 +15,9 @@ const float deg_2_rad_const = M_PI / 180.0f;
 
 /* Supply your gyro offsets here, scaled for min sensitivity */
 // Please see: https://github.com/ElectronicCats/mpu6050/blob/master/examples/IMU_Zero/IMU_Zero.ino
-const int XAccelOffset = -383;
-const int YAccelOffset = -2063;
-const int ZAccelOffset = 1559;
-const int XGyroOffset = -147;
-const int YGyroOffset = -4;
-const int ZGyroOffset = -341;
+const int16_t XAccelOffset = -383;
+const int16_t YAccelOffset = -2063;
+const int16_t ZAccelOffset = 1559;
+const int16_t XGyroOffset = -147;
+const int16_t YGyroOffset = -4;
+const int16_t ZGyroOffset = -341;
